requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
huggingface-hub>=0.17.0 